Add support for Azure OpenAI
If `aoai_endpoint` is set, then use Azure OpenAI endpoint rather than
the direct OpenAI endpoint. Requires `aoai_deployment` to be specified
as well. You should set the model too, so that aichat knows the correct
number of remaining tokens.

Also adds support for `max_tokens` since the AOAI default is very low.

The Azure OpenAI requires chats to be encoded as completions, so logic
to do this is added as well.