feat: Allow user to specify TTS voice via query param
I like the TTS feature and wanted a way to change the default voice my
browser gives me.

This change simply parses the string name of the voice you want and then
sets it when you try to use TTS. It should only run the code to set the
voice if user has requested a voice and a voice is not yet chosen.

In the case that the user sends an invalid voice name it simply will
continue to use the default voice and clear the query param for voice to
prevent a retry to set the voice. Tested this function only gets hit
once regardless of success.

I love aichat and just wanted to write this feature for myself, if you
want I am happy to make any changes or if you do not want this feature feel free to close :)