feat: implement add command (WIP / RFC)
This implements a new .add command in the repl that allows the user to
add a new client identically to how initial client/model adding works
if no client is configured / `config.yaml` is missing.

The initial prompt for adding a client is very helpful, as example clients
are pulled from (and guide the user). To subsequently add new clients
(if the user already has a configured client), the user currently has
to look up the example config file and edit yml manually. This change
would make it simpler for the end user to quickly add new clients based
on prompts rather than manually referencing the example yml.

This is WIP / RFC and will need to determine:

- Proper place to add logic, all in `src/repl/mod.rs` seems wrong;
  should this go in `src/config/mod.rs`?
- Should underlying `create_client_config` be modified to return the
  single created client?
- Related to above, could `create_client_config` be modified to not
  return Value and/or should actual types be used and then cast down
  to yml?
- Is there a way to apply config in place / apply added client without
  restarting the application?